Add depends system. Port daemon over to automake.

+++ src/test/uint160_tests.cpp
@@ -6,11 +6,11 @@ BOOST_AUTO_TEST_SUITE(uint160_tests)
 
 BOOST_AUTO_TEST_CASE(uint160_equality)
 {
-    uint160 num1 = 10;
+    uint64_t num1 = 10;
     uint160 num2 = 11;
     BOOST_CHECK(num1+1 == num2);
 
-    uint64 num3 = 10;
+    uint64_t num3 = 10;
     BOOST_CHECK(num1 == num3);
     BOOST_CHECK(num1+num2 == num3+num2);
 }
